Correct a mistake in the way overvotes are calculated
The CVR format shows no selections in the event of an overvote. To make
the audit board selections match in the event of an overvote, we
therefore have to exclude all selections when the audit board records an
overvote in order to have a chance at matching the CVR if indeed there
was an overvote.

CountyContestComparisonAudit is currently unused, but a similar change
has been made to avoid future confusion.