Add final review screen with ACVR re-upload capability
This allows for "blind" reviews of previously-audited ballots, in case
the audit board makes a mistake. It is "blind" in the sense that the
audit boards do not see what they previously entered - it is purely a
mechanism for the boards to correct a mistake that they discovered
without outside prompting before they sign off on a round.